Import Phone FFT Trained Model
Previously, we trained an SVM model based on FFT that was saved directly from the phone (instead of post-processing on laptop). Imported this model into the assets folder.
Next, copied over the conditionals fix for windowing, from the modified microphone app. This makes it so that the FFT is correctly windowed to 2-20 kHz, sampled at every 10 points.